feat: Add more locale data for new states
