+++ packages/shared/lib/typings/index.ts
@@ -4,6 +4,7 @@ import * as bridge from './bridge'
 import * as client from './client'
 import * as events from './events'
 import * as message from './message'
+import * as ledger from './ledger'
 import * as wallet from './wallet'
 
-export { account, address, bridge, client, events, message, wallet }
+export { account, address, bridge, client, events, ledger, message, wallet }
