+++ packages/shared/lib/typings/ledger.ts
@@ -1,0 +1,21 @@
+export interface LedgerAppInfo {
+    name: string,
+    version: string
+}
+
+export enum LedgerDeviceState {
+    Connected = 'connected',
+    NotDetected = 'notDetected',
+    AppNotOpen = 'appNotOpen',
+    MnemonicMismatch = 'mnemonicMismatch'
+}
+
+export interface LedgerStatusPayload {
+    type: LedgerStatus
+}
+
+export enum LedgerStatus {
+    Connected = 'Connected',
+    Disconnected = 'Disconnected',
+    Locked = 'Locked'
+}
